+++ CHANGELOG.md
@@ -4,6 +4,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- VIMEET-2205 - Display analytics on sheet
+- VIMEET-2206 - Add option to activate analytics depending on participation type
 
 ## [2.49.1] - 2020-09-25
 ### Fixed